First commit of typing game.
Current status: The game generates a new word every two seconds. Users
can type the word and add to their score. Based on the level, at a
certain score the user levels up and gets a new set of words to practice
with. If they miss a word or type it wrong, their score resets to zero.
Scoring system doesn't update very well on the form at the moment. It
also needs to only increment when the user types a word from the current
level (which it doesn't do at the moment).